requirements.txt
fastapi
python-dotenv==0.18.0
sqlalchemy
uvicorn[standard]